获取模块安装数量修改为从获取到的module实例计算
