+++ module-controller/internal/handler/modulereplicaset_replicas_changed_handler_test.go
@@ -12,5 +12,5 @@ func TestModuleReplicaSetReplicasChangedHandler(t *testing.T) {
 }
 
 type TestModuleReplicaSetReplicasChangedClient struct {
-	utils.MockOnBaseClient
+	utils.MockClient
 }
