+++ module-controller/internal/utils/test_utils_test.go
@@ -1,7 +1,9 @@
 package utils
 
 import (
+	"github.com/sofastack/sofa-serverless/api/v1alpha1"
 	"github.com/stretchr/testify/assert"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"testing"
 )
@@ -10,6 +12,12 @@ func TestPrepareModuleDeployment(t *testing.T) {
 	moduleDeploymentName := "testModuleDeploymentName"
 	moduleDeployment := PrepareModuleDeployment("default", moduleDeploymentName)
 	assert.Equal(t, moduleDeploymentName, moduleDeployment.Name)
+}
+
+func TestPrepareModuleReplicaSet(t *testing.T) {
+	moduleReplicaSetName := "testModuleReplicaSet"
+	moduleReplicaSet := PrepareModuleReplicaSet("default", moduleReplicaSetName)
+	assert.Equal(t, moduleReplicaSetName, moduleReplicaSet.Name)
 }
 
 func TestMockClient(t *testing.T) {
@@ -29,4 +37,8 @@ func TestMockClient(t *testing.T) {
 	assert.Equal(t, nil, mockClient.DeleteAllOf(nil, nil))
 	assert.Equal(t, nil, mockClient.Status())
 	assert.Equal(t, nil, mockClient.SubResource(""))
+
+	assert.Equal(t, nil, mockClient.List(nil, &v1alpha1.ModuleList{}))
+	assert.Equal(t, nil, mockClient.List(nil, &corev1.PodList{}))
+
 }
